requirements.txt
pyrogram
TgCrypto
telethon
requests
pillow
numpy
aiohttp
beautifulsoup4
fake_useragent
